adjusted to not delete when has link

+++ src/main/java/com/meiorganizadinho/service/ClientService.java
@@ -69,6 +69,10 @@ public class ClientService {
         Client client = clientRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException((("Cliente não encontrado"))));
 
+        int qtdAppointments = client.getAppointments().size();
+        if(qtdAppointments > 0) {
+            throw new BusinessException("Cliente possui vínculos com " + qtdAppointments + " appointment(s)");
+        }
         clientRepository.delete(client);
     }
 }
